Completed implementation of service initialization and removal.
Filled out servicectl (renamed from systemctl) operations to create and
remove operations including complete SQL scripts as well, proper error
logging in systemctl.log, and properly emiting output to the console.

+++ services/goldfin-api-server/src/main/java/io/goldfin/invoice/core/ProgressReporter.java
@@ -3,11 +3,9 @@
  */
 package io.goldfin.invoice.core;
 
-import java.math.BigDecimal;
-
 /**
  * Denotes a class that receives progress updates. 
  */
 public interface ProgressReporter {
-	public void progress(String description, BigDecimal percent);
+	public void progress(String description, double percent);
 }
